+++ resources/install_imagegw.rb
@@ -10,6 +10,7 @@ property :git_repo, String, default: lazy { shifter_git_repo }
 property :version, String, default: lazy { shifter_version_stable }
 property :shifter_etc_files, String, default: lazy { shifter_etc_files_dir }
 property :image_path, String, default: lazy { shifter_image_dir }
+property :expand_path, String, default: lazy { shifter_expand_dir }
 property :imagegw_fqdn, String, default: ''
 
 action :install do
@@ -26,12 +27,80 @@ action :install do
     not_if { ::File.exist?(new_resource.udiroot + '/bin/shifterimg') }
   end
 
+  shifter_imagegw_packages.each(&method(:package))
+
+  directory "#{new_resource.image_path}/cache" do
+    owner 'root'
+    group 'root'
+    mode '0755'
+    recursive true
+    action :create
+    not_if { ::File.directory?("#{new_resource.image_path}/cache") }
+  end
+
+  directory new_resource.expand_path do
+    owner 'root'
+    group 'root'
+    mode '0755'
+    recursive true
+    action :create
+    not_if { ::File.directory?(new_resource.expand_path) }
+  end
+
+  template "#{new_resource.config_dir}/imagemanager.json" do
+    source 'imagemanager.json.erb'
+    cookbook new_resource.cookbook
+    variables(
+        image_dir: new_resource.image_path,
+        expand_dir: new_resource.expand_path,
+        cache_dir: "#{new_resource.image_path}/cache",
+        system_name: system_name
+    )
+  end
+
+  template "#{systemd_service_dir}/shifter_imagegw.service" do
+    source 'shifter_imagegw.service.erb'
+    cookbook new_resource.cookbook
+    variables(
+        shifter_udiroot_dir: shifter_udiroot_dir
+    )
+  end
+
+  execute 'Systemd Daemon Reload' do
+    command 'systemctl daemon-reload'
+    action :run
+  end
+
   link "#{new_resource.udiroot}/bin/shifterimg" do
     to '/usr/bin/shifterimg'
+  end
+
+  service 'Shifter Image Manager Service' do
+    service_name 'shifter_imagegw'
+    supports restart: true, status: true
+    restart_command "systemctl restart #{service_name}"
+    status_command "systemctl status #{service_name}"
+    action :start
   end
 end
 
 action :uninstall do
+  service 'Shifter Image Manager Service' do
+    service_name 'shifter_imagegw'
+    supports restart: true, status: true
+    restart_command "systemctl restart #{service_name}"
+    status_command "systemctl status #{service_name}"
+    action :stop
+  end
+
+  file '#{systemd_service_dir}/shifter_imagegw.service' do
+    action :delete
+  end
+
+  file '#{new_resource.config_dir}/imagemanager.json' do
+    action :delete
+  end
+
   link "#{new_resource.udiroot}/bin/shifterimg" do
     to '/usr/bin/shifterimg'
     action :delete